Code/tweet_gen_app/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
entrypoints==0.3
flake8==3.7.8
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==19.9.0
idna==2.8
image==1.5.27
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
oauthlib==3.1.0
packaging==19.1
Pillow==6.2.1
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pydocstyle==4.0.0
pyfiglet==0.8.post1
pyflakes==2.1.1
pymongo==3.9.0
pypandoc==1.4
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
randomwordgenerator==1.0.0
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
snowballstemmer==1.9.0
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.2
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
